requirements.txt
flask
flask_restful
flask_cors
python-escpos